# Speed up `make_race_combo`